moving text sizing functions up to
