+++ py_src/ipyelk/nx/nx.py
@@ -1,39 +1,12 @@
 # Copyright (c) 2020 Dane Freeman.
 # Distributed under the terms of the Modified BSD License.
-from dataclasses import dataclass
 from itertools import tee
 from typing import Dict, Hashable, List, Optional, Tuple
 
 import networkx as nx
 
-from ..diagram.elk_model import ElkNode, ElkPort, ElkRoot
-
-
-@dataclass(frozen=True)
-class Edge:
-    source: Hashable
-    source_port: Optional[Hashable]
-    target: Hashable
-    target_port: Optional[Hashable]
-    owner: Hashable
-    data: Optional[Dict]
-
-    def __hash__(self):
-        return hash((self.source, self.source_port, self.target, self.target_port))
-
-
-@dataclass(frozen=True)
-class Port:
-    node: Hashable
-    elkport: ElkPort
-
-    def __hash__(self):
-        return hash(tuple([hash(self.node), hash(self.elkport.id)]))
-
-
-NodeMap = Dict[Hashable, ElkNode]
-EdgeMap = Dict[Hashable, List[Edge]]
-PortMap = Dict[Hashable, Port]
+from ..diagram.elk_model import ElkNode, ElkRoot
+from ..transform import NodeMap
 
 
 def compact(array: Optional[List]) -> Optional[List]:
@@ -107,3 +80,54 @@ def get_ports(edge_data: Dict) -> Tuple[Optional[Hashable], Optional[Hashable]]:
     source_port = edge_data.get("sourcePort", None) or p
     target_port = edge_data.get("targetPort", None) or p
     return source_port, target_port
+
+
+def is_hidden(tree: nx.DiGraph, node: Hashable, attr: str) -> bool:
+    """Iterate  on the node ancestors and determine if it is hidden along the chain"""
+    if tree and node in tree:
+        if tree.nodes[node].get(attr, False):
+            return True
+        for ancestor in nx.ancestors(tree, node):
+            if tree.nodes[ancestor].get(attr, False):
+                return True
+    return False
+
+
+def build_hierarchy(
+    g: nx.Graph, tree: nx.DiGraph, elknodes: NodeMap, HIDDEN_ATTR: str
+) -> List[ElkNode]:
+    """The Elk JSON is hierarchical. This method iterates through the build
+    elknodes and links children to parents if the incoming source includes a
+    hierarcichal networkx diagraph tree.
+
+    :param g: [description]
+    :type g: nx.Graph
+    :param tree: [description]
+    :type tree: nx.DiGraph
+    :param elknodes: mapping of networkx nodes to their elknode representations
+    :type elknodes: NodeMap
+    :param HIDDEN_ATTR: [description]
+    :type HIDDEN_ATTR: str
+    :return: Top level ElkNodes to put as children under the ElkRoot
+    :rtype: List[ElkNode]
+    """
+    if tree:
+        # roots of the tree
+        roots = [n for n, d in tree.in_degree() if d == 0]
+        for n, elknode in elknodes.items():
+            if n in tree:
+                elknode.children = [
+                    elknodes[c]
+                    for c in tree.neighbors(n)
+                    if not is_hidden(tree, c, HIDDEN_ATTR)
+                ]
+            else:
+                # nodes that are not in the tree
+                roots.append(n)
+    else:
+        # only flat graph provided
+        roots = []
+        for n, elknode in elknodes.items():
+            if not is_hidden(tree, n, HIDDEN_ATTR):
+                roots.append(n)
+    return [elknodes[n] for n in roots]
